Revert "Using classes, and tweaking sided jumps"
This reverts commit 6ac82009

+++ src/CharactersServer/Characters/WolverineServer.h
@@ -14,8 +14,8 @@ public:
     void makeBuilderStruct(character_builder_t* builder, bool isFirstTeam);
 
 private:
-    void moveRight(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
-    void moveLeft(int distance, float vel, Box *boxOfEnemy, bool isGrounded);
+    void moveRight(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
+    void moveLeft(int distance, int vel, Box *boxOfEnemy, bool isGrounded);
     void stand();
 
     int getSpriteNumber();
